src/backend/pyproject.toml
[project]
name = "hackathon-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "redis>=5.2",
    "neo4j>=5.26",
    "chromadb-client>=0.6",
    "bcrypt>=4.2",
    "pyjwt>=2.10",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "structlog>=24.4",
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "openai-agents>=0.1",
    "anthropic>=0.40",
    "websockets>=14.0",
    "python-multipart>=0.0.12",
    "psycopg2-binary>=2.9",
    "duckdb>=1.1",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "sse-starlette>=2.0",
    "email-validator>=2.3.0",
    "strawberry-graphql[fastapi]>=0.254",
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50b",
    "opentelemetry-instrumentation-httpx>=0.50b",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b",
    "opentelemetry-instrumentation-redis>=0.50b",
    "mangum>=0.19",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.uv.workspace]
members = ["."]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
